+++ ui/litellm-dashboard/src/app/(dashboard)/policies/_components/guardrail_selection_modal.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { renderWithProviders } from "../../../tests/test-utils";
+import { renderWithProviders } from "@/../tests/test-utils";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import GuardrailSelectionModal from "./guardrail_selection_modal";
