+++ ui/litellm-dashboard/src/app/(dashboard)/policies/_components/impact_popover.test.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { renderWithProviders } from "../../../tests/test-utils";
+import { renderWithProviders } from "@/../tests/test-utils";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import * as networking from "../networking";
+import * as networking from "@/components/networking";
 import ImpactPopover from "./impact_popover";
-import { PolicyAttachment } from "./types";
+import { PolicyAttachment } from "@/components/policies/types";
 
-vi.mock("../networking");
+vi.mock("@/components/networking");
 
 vi.mock("@heroicons/react/outline", () => ({
   EyeIcon: function EyeIcon() {
