+++ ui/litellm-dashboard/src/app/(dashboard)/policies/_components/policy_templates.tsx
@@ -8,7 +8,7 @@ import {
   CurrencyDollarIcon,
   CheckCircleIcon,
 } from "@heroicons/react/outline";
-import { getPolicyTemplates } from "../networking";
+import { getPolicyTemplates } from "@/components/networking";
 
 interface PolicyTemplateCardProps {
   title: string;
